requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-cors-headers==1.1.0
django-filter==0.11.0
django-herokuapp==0.9.20
django-require==1.0.8
django-require-s3==1.0.0
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.2.4
gunicorn==19.3.0
Markdown==2.6.2
Pillow==3.0.0
psycopg2==2.6.1
pytz==2015.6
sh==1.11
source==1.2.0
static3==0.6.1
virtualenv==13.1.2
waitress==0.8.10
wheel==0.26.0
